Use AbstractQuery & Query division consistently
